// supabase/functions/_shared/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      horario: {
        Row: {
          anio_fin: number
          anio_inicio: number
          cod_carrera: number
          cod_periodo: string | null
          created_at: string
          id_horario: number
          pao: number
          paralelo: number
        }
        Insert: {
          anio_fin: number
          anio_inicio: number
          cod_carrera: number
          cod_periodo?: string | null
          created_at?: string
          id_horario?: number
          pao: number
          paralelo: number
        }
        Update: {
          anio_fin?: number
          anio_inicio?: number
          cod_carrera?: number
          cod_periodo?: string | null
          created_at?: string
          id_horario?: number
          pao?: number
          paralelo?: number
        }
        Relationships: [
          {
            foreignKeyName: "horario_cod_carrera_fkey"
            columns: ["cod_carrera"]
            isOneToOne: false
            referencedRelation: "carrera"
            referencedColumns: ["id_carrera"]
          },
        ]
      }
      horario_materia: {
        Row: {
          cod_horario: number | null
          cod_materia: number
          created_at: string
          hora_fin: number
          hora_inicio: number
          id_horario_materia: number
          numero_dia: number
          paralelo: number
        }
        Insert: {
          cod_horario?: number | null
          cod_materia: number
          created_at?: string
          hora_fin: number
          hora_inicio: number
          id_horario_materia?: number
          numero_dia: number
          paralelo: number
        }
        Update: {
          cod_horario?: number | null
          cod_materia?: number
          created_at?: string
          hora_fin?: number
          hora_inicio?: number
          id_horario_materia?: number
          numero_dia?: number
          paralelo?: number
        }
        Relationships: [
          {
            foreignKeyName: "horario_materia_cod_horario_fkey"
            columns: ["cod_horario"]
            isOneToOne: false
            referencedRelation: "horario"
            referencedColumns: ["id_horario"]
          },
          {
            foreignKeyName: "horario_materia_cod_materia_fkey"
            columns: ["cod_materia"]
            isOneToOne: false
            referencedRelation: "materia"
            referencedColumns: ["id_materia"]
          },
        ]
      }
// ...
      materia: {
        Row: {
          cod_carrera: number
          cod_docente: number | null
          cod_malla: number | null
          codigo: string
          created_at: string
          id_materia: number
          nombre: string
          numero_creditos: number
          pao: number
        }
        Insert: {
          cod_carrera: number
          cod_docente?: number | null
          cod_malla?: number | null
          codigo: string
          created_at?: string
          id_materia?: number
          nombre: string
          numero_creditos: number
          pao: number
        }
        Update: {
          cod_carrera?: number
          cod_docente?: number | null
          cod_malla?: number | null
          codigo?: string
          created_at?: string
          id_materia?: number
          nombre?: string
          numero_creditos?: number
          pao?: number
        }
        Relationships: [
          {
            foreignKeyName: "materia_cod_carrera_fkey"
            columns: ["cod_carrera"]
            isOneToOne: false
            referencedRelation: "carrera"
            referencedColumns: ["id_carrera"]
          },
          {
            foreignKeyName: "materia_cod_docente_fkey"
            columns: ["cod_docente"]
            isOneToOne: false
            referencedRelation: "docente"
            referencedColumns: ["id_docente"]
          },
        ]
      }
// ...
    }
// ...
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never

export type TablesInsert<
  PublicTableNameOrOptions extends
    | keyof PublicSchema["Tables"]
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? Database[PublicTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : PublicTableNameOrOptions extends keyof PublicSchema["Tables"]
    ? PublicSchema["Tables"][PublicTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never